feat(web): move workspace-zoom control into the transport footer
The workspace-zoom cluster (ZOOM - nn% +) was a floating pill at the
bottom-right of the workspace, beside the Add Panel button. Move it into
the bottom transport bar next to the PA-temperature chip so workspace-level
status reads together, and restyle it from a floating pill to the footer
.chip vocabulary (dark inset, hairline border).

Extract WorkspaceZoomControls into its own component. Behaviour is unchanged:
optimistic step + POST /api/ui/workspace-zoom + applyState reconcile, with the
percent readout doubling as a click-to-reset-to-100% button.
